Refactored screens logic for the new state machine
