+++ app/src/main/java/ru/whoame/recogniser/ui/RecogniserApp.kt
@@ -5,19 +5,12 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import org.koin.androidx.compose.koinViewModel
 import ru.whoame.recogniser.ui.screen.recognisedlist.RecognisedListScreen
 
 @Composable
 @OptIn(ExperimentalMaterial3AdaptiveApi::class)
-fun RecogniserApp(
-    viewModel: RecogniserAppViewModel = koinViewModel(),
-) {
-    val name by viewModel.stateFlow.collectAsStateWithLifecycle()
-
+fun RecogniserApp() {
     Scaffold { innerPaddings ->
         val screenModifier = Modifier
             .fillMaxSize()
