+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/factory/DeleteRecognisedItemDefaultsFactory.kt
@@ -3,7 +3,7 @@ package ru.whoame.recogniser.ui.dialog.factory
 import ru.whoame.recogniser.model.RecognisedObject
 import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemDomainEvent
 import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemState
-import ru.whoame.recogniser.ui.viewmodel.contracts.BaseDefaultsFactory
+import ru.whoame.state_machine.contract.BaseDefaultsFactory
 
 class DeleteRecognisedItemDefaultsFactory(
     private val item: RecognisedObject,
