+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/factory/DeleteRecognisedItemStateFactory.kt
@@ -2,8 +2,8 @@ package ru.whoame.recogniser.ui.dialog.factory
 
 import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemState
 import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemUiState
-import ru.whoame.recogniser.ui.viewmodel.contracts.BaseStateFactory
 import ru.whoame.recogniser.utils.isLoading
+import ru.whoame.state_machine.contract.BaseStateFactory
 
 class DeleteRecognisedItemStateFactory : BaseStateFactory<DeleteRecognisedItemState, DeleteRecognisedItemUiState> {
