+++ app/src/main/java/ru/whoame/recogniser/ui/screen/recognisedlist/RecognisedListViewModel.kt
@@ -1,11 +1,11 @@
 package ru.whoame.recogniser.ui.screen.recognisedlist
 
+import ru.whoame.recogniser.ui.base.StateMachineViewModel
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.factory.ListScreenDefaultsFactory
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.factory.ListScreenStateFactory
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.handler.ListScreenDomainEventHandler
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.handler.ListScreenUiEventHandler
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.model.*
-import ru.whoame.recogniser.ui.viewmodel.StateMachineViewModel
 
 class RecognisedListViewModel(
     defaultsFactory: ListScreenDefaultsFactory,
