+++ app/src/main/java/ru/whoame/recogniser/ui/screen/recognisedlist/state/handler/ListScreenDomainEventHandler.kt
@@ -6,7 +6,7 @@ import ru.whoame.recogniser.data.repository.RecognisedObjectRepository
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.model.ListScreenDomainEvent
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.model.ListScreenSideEffect
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.model.ListScreenState
-import ru.whoame.recogniser.ui.viewmodel.handler.BaseDomainEventHandler
+import ru.whoame.state_machine.handler.BaseDomainEventHandler
 
 class ListScreenDomainEventHandler(
     private val repository: RecognisedObjectRepository,
