+++ app/src/main/java/ru/whoame/recogniser/ui/screen/recognisedlist/state/model/ListScreenSideEffect.kt
@@ -1,7 +1,9 @@
 package ru.whoame.recogniser.ui.screen.recognisedlist.state.model
 
+import ru.whoame.recogniser.model.RecognisedObject
+
 sealed interface ListScreenSideEffect {
 
-    data class DeleteItem(val id: Long) : ListScreenSideEffect
+    data class DeleteItem(val item: RecognisedObject) : ListScreenSideEffect
 
 }
